Refactor sphinx_ext into submodules
